Changed empty dashboard components to group file
