+++ frontend/src/modules/Dashboard/Styles/Groups.style.tsx
@@ -1,5 +1,8 @@
 import styled from 'styled-components'
-import { GenericHTMLProps, h2, StyledComponent } from '@core'
+import { ReactComponent as NoClassesIcon } from '@assets/img/noclassesicon.svg'
+import { GenericHTMLProps, h2, StyledComponent, h3, h4 } from '@core'
+
+const NoClasses = NoClassesIcon
 
 export const StyledContainer = styled.div`
   display: flex;
@@ -33,4 +36,34 @@ export const StyledText = styled.div`
   ${h2};
   font-weight: 300;
   margin-top: 1rem;
+`
+
+export const StyledTextBox = styled.div`
+  ${h4};
+  text-align: center;
+`
+
+export const StyledSmallText = styled.div`
+  ${h3};
+  font-size: 1.3rem !important;
+  width: 50%;
+  height: 30%;
+  word-wrap: break-word;
+  margin: auto;
+  text-align: center;
+`
+export const StyledClassesContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  padding: 5.5rem;
+  flex-direction: row;
+  flex-wrap: wrap;
+`
+export const StyledNoClasses = styled(NoClasses)`
+  position: static;
+  width: 295px;
+  height: 300px;
+  left: 0px;
+  top: 0px;
+  padding: 15px;
 `
